Rename to libcyphal++.
