+++ extras/cyphal++/examples/example-01-opencyphal-basic-node/example-01-opencyphal-basic-node.cpp
@@ -17,7 +17,7 @@
 #include <atomic>
 #include <iostream>
 
-#include <cyphalcan++/cyphalcan++.h>
+#include <cyphal++/cyphal++.h>
 
 /**************************************************************************************
  * CONSTANT
